fix(dependency_injection): Test generated from mason brick cause compilation errors
