+++ bricks/dependency_injection/__brick__/lib/app/di/data/annotations.dart
