+++ bricks/dependency_injection/__brick__/lib/app/di/data/config/app_config.dart
@@ -1,18 +1,13 @@
 // coverage:ignore-file
+// This file uses a static function. We cannot test it directly because it is
+// not a class.
 
-import 'package:{{project_name}}/app/di/helper/config/app_config.config.dart';
-import 'package:{{project_name}}/app/di/helper/config/get_it_config.dart';
+import 'package:{{project_name}}/app/di/data/config/app_config.config.dart';
+import 'package:{{project_name}}/app/di/data/config/get_it_config.dart';
 import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
 
-
-/// Sets up all the pieces your app needs to run.
-///
-/// Think of this as the main setup coordinator for your app. It's like
-/// a checklist that makes sure everything is ready before the app starts:
-/// - Gets the right settings for your environment (dev, staging, prod)
-/// - Sets up all services (like Firebase, authentication, etc.)
-/// - Makes sure everything is connected properly
+/// Uses GetIt to set up all the pieces your app needs to run.
 class GetItAppConfig implements GetItConfig {
   /// Starts the setup process for all app dependencies.
   ///
@@ -24,16 +19,16 @@ class GetItAppConfig implements GetItConfig {
   /// It uses _configureAppDependencies to do the actual work.
   @override
   Future<void> configureDependencies(
-      String environment,
-      GetIt getIt,
-      ) =>
+    String environment,
+    GetIt getIt,
+  ) =>
       _configureAppDependencies(getIt, environment);
 }
 
 /// The function that actually sets up all app dependencies.
 ///
-/// @InjectableInit is like a magic wand that:
-/// - Looks through your code for special annotations (@injectable,
+/// @InjectableInit is a special annotation that tells the app to:
+/// - Look through your code for other special annotations (@injectable,
 ///   @singleton, etc.)
 /// - Creates all the necessary services automatically
 /// - Connects everything together in the right order
@@ -44,7 +39,7 @@ class GetItAppConfig implements GetItConfig {
 /// only GetItAppConfig can use it. This helps keep our code organized and safe.
 @InjectableInit(initializerName: 'init')
 Future<void> _configureAppDependencies(
-    GetIt getIt,
-    String environment,
-    ) async =>
+  GetIt getIt,
+  String environment,
+) async =>
     getIt.init(environment: environment);
